fix norme

+++ libft/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: opodolia <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/01 14:26:56 by opodolia          #+#    #+#             */
-/*   Updated: 2017/03/12 15:00:39 by opodolia         ###   ########.fr       */
+/*   Updated: 2017/03/13 13:08:27 by opodolia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t len)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (i <= len)
